Testing and fixes (#4)
* Fix typo in class name

* Add some basic tests for message parsing

Add tests for successful message parsing (when there is 1 key) and
another for unsuccessful (when there is more than 1 key).

* Add another test for empty messages

Add a test that checks that empty messages will fail to parse.

* Create `sybl.json` and directories if not found

Previously, Mallus would crash if `sybl.json` did not exist. Instead,
this will check whether the directory exists and create it and its
parents if not before ensuring `sybl.json` exists.

Fixes #1.