Add typechecking through MyPy
Add typechecking through the use of MyPy on all files. This currently
passes and required some updates to annotations and method calls to
work. MyPy seems to be pretty strict, but it should reduce the number of
errors in the future once we are used to it.

Update some docstrings and Pylint warnings to bring the score closer to
10. The last couple of issues surround the `JobConfig` handling which is
currently unimplemented, so it's not worth fixing them for now.