feat: implement more complex config handling
Firstly, change job configurations to expect
`message_creation_timestamp` and `prediction_cutoff_timestamp` fields,
which inform the client of when the DCL processed the message and when
it expects predictions to be returned by. Sending the creation time
gives a better idea of the latency between the two and allows the client
to calculate the estimated time of the job. Sending the cutoff time
provides a fixed point in time at which the client should return by,
meaning they can stop early if processing is taking longer than expected
without being penalised.

Additionally, allow users to specify a custom function for handling job
configurations. This allows for more advanced usage of the data provided
to them, such as using different timeouts for different job types as
shown in the unit tests.