fix: find the correct `.env` file
When searching for the `.env` file, `find_dotenv()` seems to use the
directory of `mallus` itself instead of the current directory when
installed locally. This explains why it finds a `.env` file, as it is
looking for the one I have within `mallus` itself.

Instead, use `pathlib` to get the current execution directory and use
the `.env` file there instead.